refactor(Product.java): 상품 모델을 entity 로 설정
- ProductPrice 상품 가격 모델 구현 및 테스트 작성

+++ src/main/java/kitchenpos/dto/ProductResponse.java
@@ -47,8 +47,9 @@ public class ProductResponse {
             return false;
         }
         ProductResponse that = (ProductResponse) o;
-        return Objects.equals(id, that.id) && Objects.equals(name, that.name)
-                && Objects.equals(price, that.price);
+        return Objects.equals(id, that.id) &&
+                Objects.equals(name, that.name) &&
+                Objects.equals(price.longValue(), that.price.longValue());
     }
 
     @Override
